Commit `checkUseBeforeDeclare` code
Currently untested, but it's pretty beautiful as it is.
Needs to be abstracted away some more.

The function `checkDupe` needs to be helper-ized too.

Functions `dupeParam` and `dupeVariable` will become constructors of the `Error`
type, when it stops beign a `newtype String`

This is, all in all, much better than the previous version, though it only
registers ATM instead of also checking.